requirements.txt
pillow
httpx
google-genai
google-adk
python-dotenv